requirements.txt
tqdm
torchinfo
h5py
einops
numpy>=1.24.3,<2.0
matplotlib>=3.7.5
pandas>=2.0.3
arch>=5.3.1
hurst>=0.0.5
tslearn>=0.6.3
cython>=3.0.10
scikit-learn>=1.3.2
stumpy>=1.12.0
networkx>=3.1
transformers>=4.38.0
torch==2.3.0
chronos-forecasting>=2.0